Commit build updates for SLES/OpenSUSE

+++ build/linux/setup.sh
@@ -5,6 +5,7 @@ PYTHONTAR="Python-2.7.14"
 PYTHONVER="python2.7"
 PYTHONBIN="/usr/local/bin/python2.7"
 CXFREEZEVER="cx_Freeze-4.3.4"
+SKIP_PYTHON=0
 
 # Get information about system
 . $BUILD_DIR/linux/init.sh
@@ -22,13 +23,53 @@ install_prereqs() {
 
 
     if [ "$distro" == "Debian" ] || [ "$distro" == "Ubuntu" ]; then
-        apt-get install debian-builder rpm gcc gcc-c++ git wget openssl libssl-dev libffi-dev sqlite3 libsqlite3-dev zlib1g-dev alien -y
+    
+        apt-get install debian-builder rpm gcc gcc-c++ wget openssl libssl-dev libffi-dev sqlite3 libsqlite3-dev zlib1g-dev alien -y
+    
     elif [ "$distro" == "CentOS" ] || [ "$distro" == "RHEL" ] || [ "$distro" == "Oracle" ] || [ "$distro" == "CloudLinux" ]; then
+    
         yum install epel-release -y
         yum install gcc gcc-c++ zlib zlib-devel openssl openssl-devel rpm-build libffi-devel sqlite sqlite-devel wget -y
-    elif [ "$distro" == "SUSE LINUX" ]; then
-        zypper isntall gcc
+
+    elif [ "$distro" == "SUSE LINUX" ] || [ "$distro" == "SLES" ]; then
+
+        # We don't need to install python on SLES due to their updated version of python
+        # available with the OS itself
+        if [ "$dist" == "sles15" ] || [ "$dist" == "sles12" ] || [ "$distro" == "OpenSUSE" ]; then
+
+            zypper install gcc gcc-c++ python python-devel zlib zlib-devel openssl libopenssl-devel sqlite3 sqlite3-devel rpm-build wget
+            SKIP_PYTHON=1
+            PYTHONBIN=$(which python2.7)
+
+        elif [ "$dist" == "sles11" ]; then
+
+            # Show information about suse manual config for SDK
+            echo ""
+            echo "Manual configuration required:"
+            echo ""
+            echo "1. Register your SUSE system."
+            echo "2. Install SDK repo:"
+            echo "   https://www.suse.com/support/kb/doc/?id=7015337"
+            if [ "$arch" == "i686" ]; then
+                $arch="i586";
+            fi
+            echo "   https://nu.novell.com/repo/\$RCE/SLE11-SDK-SP4-Pool/sle-11-$arch/rpm/$arch/sle-sdk-release-11.4-1.55.$arch.rpm"
+            echo "   https://nu.novell.com/repo/\$RCE/SLE11-SDK-SP4-Pool/sle-11-$arch/rpm/$arch/sle-sdk-release-SDK-11.4-1.55.$arch.rpm"
+            echo "3. Install the Security Module:"
+            echo "   https://www.suse.com/documentation/suse-best-practices/singlehtml/securitymodule/securitymodule.html"
+            echo ""
+            echo "Press enter to continue..."
+            read
+
+            zypper install gcc gcc-c++ zlib zlib-devel sqlite3 sqlite3-devel rpm-build wget
+
+            # Install openssl 1.0.x for TLS 1.2 support
+            zypper install openssl1 libopenssl1_0_0 libopenssl1-devel libcurl4-openssl1 curl-openssl1 wget-openssl1
+
+        fi
+
     else
+
         echo "Could not determine your OS type... you need to install the following dependencies:"
         echo ""
         echo "gcc, gcc-c++"
@@ -43,6 +84,7 @@ install_prereqs() {
         echo "If you're running a debian distro you must also install debian-builder"
         echo ""
         exit 1
+
     fi
 
 
@@ -54,13 +96,15 @@ install_prereqs() {
     cd $BUILD_DIR/resources
 
     # Install bundled Python version from source
-    tar xf $PYTHONTAR.tgz
-    cd $PYTHONTAR
-    ./configure --with-zlib=/usr/include --enable-shared && make && make altinstall
-    echo '/usr/local/lib' >> /etc/ld.so.conf 
-    /sbin/ldconfig
-    cd ..
-    rm -rf $PYTHONTAR
+    if [ $SKIP_PYTHON -eq 0 ]; then
+        tar xf $PYTHONTAR.tgz
+        cd $PYTHONTAR
+        ./configure --with-zlib=/usr/include --enable-shared && make && make altinstall
+        echo '/usr/local/lib' >> /etc/ld.so.conf 
+        /sbin/ldconfig
+        cd ..
+        rm -rf $PYTHONTAR
+    fi
 
     # Install the patched version of cx_Freeze
     tar xf $CXFREEZEVER.tar.gz
